+++ lizard_portal/templates/portals/communique.js
@@ -43,7 +43,7 @@
                 });
            }{% if user.is_authenticated %},
             tools: [{
-                type: 'save',
+                type: 'edit',
                 handler: function(e, target, panelHeader, tool){
                     console.log(arguments)
                     var portlet = panelHeader.up('panel')