+++ lizard_portal/templates/portals/krw-overzicht.js
@@ -54,7 +54,7 @@
                 }
             }),
             tools: [{
-                type: 'save',
+                type: 'edit',
                 handler: function (e, target, panelHeader, tool) {
                     var cm = Ext.getCmp('portalWindow').context_manager.getContext();