+++ lizard_portal/templates/portals/toestand-aan-afvoergebied.js
@@ -52,7 +52,7 @@
                 }
             }),
             tools: [{
-                type: 'save',
+                type: 'edit',
                 handler: function (e, target, panelHeader, tool) {
                     var cm = Ext.getCmp('portalWindow').context_manager.getContext();